Added the version 4 spec PDF. IT was hard to find
This document show that my boy Sam Nicaise ordered 9 MPPT. It also show
the specs that these guys came with.

The most critical part for us is the can conversion Factors. We will use
this number to re calculate our upper and lower limits for the trackers
when we re-build our code.

Drive Tek changes the hardware depending on the use of the product. This
document is important because we can use it to understand critical aspecs
of the hardware.

Brooke is gone to AFROTC...